Adding config script
Summary:
Adding a config script equivalent to those from tf and addons.
The script sets up the venv to contain TensorFlow, and the .bazelrc.user to setup the environment for Poplar.
The main difference from the addons and tf scripts are that they are bash scripts and this is a Python script.

Test Plan: From local testing I've confirmed that TensorFlow and Poplar are correctly included for builds and tests, though nothing can build to completion yet because tf hasn't been updated to 2.6.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, yanislavd, #tensorflow

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, yanislavd, #tensorflow

Subscribers: georgep

Maniphest Tasks: T57260

Differential Revision: https://phabricator.sourcevertex.net/D63145