Edit Keras docs introduction lines that mention TensorFLow
Summary:
TF2.6 only

This diff was originally served as a subtask of [T60648](https://phabricator.sourcevertex.net/T60648). The original intention was to change the following line
```
The Graphcore implementation of TensorFlow includes Keras support for IPUs.
```

into

```
The Graphcore implementation of TensorFlow2 includes Keras support for IPUs.
```

so that we make it explicit that the Keras APIs here are for TF2.

Now that Keras is moved out from TensorFlow as a separate package, this line still needs to be edited.

Reviewers: #tensorflow, #documentation, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jamiep

Reviewed By: #tensorflow, #documentation, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jamiep

Subscribers: grahamh

Maniphest Tasks: T60648

Differential Revision: https://phabricator.sourcevertex.net/D68776