Fix exporting model for serving with unspecified batch size
Summary:
In the previous implementation user was not able to set a batch size value of the model just by passing the `batch_size` argument to the serving API's method.
The check for `None` in signature shape was done to early - before the batch size was overridden by the value passed by the user.

Ref T62150

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T62150

Differential Revision: https://phabricator.sourcevertex.net/D70768